# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gallery API",
        "description": "Kenji's Gallery API",
        "contact": {
            "email": "N00202534@iadt.ie"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0.0"
    },
    "paths": {
        "/api/art": {
            "get": {
                "tags": [
                    "Art"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Displays all the Art pieces",
                "operationId": "02497dfe4b84b4200a4856599cd76bff",
                "responses": {
                    "200": {
                        "description": "Successful operation, Returns a list of Gallery in JSON format"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Art"
                ],
                "summary": "Create a new art piece",
                "description": "Stores the Art in the DB",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "category",
                                    "description",
                                    "artists",
                                    "likes"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample Title"
                                    },
                                    "category": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Autobiography"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "A long description about this great A"
                                    },
                                    "artist": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Me"
                                    },
                                    "likes": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/art/{id}": {
            "get": {
                "tags": [
                    "Art"
                ],
                "summary": "Display the specified resource.",
                "description": "Gets an Art Piece by ID",
                "operationId": "71b281c56d64848b7511ffe1deae02bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Art id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Art"
                ],
                "summary": "Delete an Art Piece",
                "description": "Delete Art",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a Art piece",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "204"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/artists": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "description": "Displays all the artist",
                "operationId": "2c8aed1387cca1c65d98da7ee11ac926",
                "responses": {
                    "200": {
                        "description": "Successful operation, Returns a list of Galleries in JSON format"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Artists"
                ],
                "summary": "Create a new artist",
                "description": "Stores the Art in the DB",
                "operationId": "83d2888bdc6fac66b9a7cc81c4c950fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "artists"
                                ],
                                "properties": {
                                    "artist": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Me"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/artists/{id}": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Display the specified resource.",
                "description": "Gets an Artist by ID",
                "operationId": "d0b68c06b8bc310fbe24c6d3ea80dac3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artist id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Artists"
                ],
                "summary": "Delete an Artist",
                "description": "Delete Artist",
                "operationId": "4f6d8bd2c8b4b738af7589b72348ac20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a Artist",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "204"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}